requirements.txt
# Core bot framework
aiogram==3.7.0
magic-filter==1.0.12

# Database
tortoise-orm==0.21.4
aerich==0.7.2
aiosqlite==0.20.0
asyncpg==0.29.0

# State management
redis==5.0.7

# Configuration
environs[dotenv]==11.0.0
